refactor: move changelog ui to new module changelog:internal
