+++ app-k9mail/build.gradle.kts
@@ -180,6 +180,7 @@ dependencies {
     testImplementation(projects.plugins.openpgpApiLib.openpgpApi)
     testImplementation(libs.appauth)
     testImplementation(projects.feature.changelog.api)
+    testImplementation(projects.feature.changelog.internal)
 }
 
 dependencyGuard {
