+++ app-k9mail/src/test/kotlin/app/k9mail/DependencyInjectionTest.kt
@@ -18,7 +18,6 @@ import com.fsck.k9.job.SyncDebugWorker
 import com.fsck.k9.mailstore.AttachmentResolver
 import com.fsck.k9.message.html.DisplayHtml
 import com.fsck.k9.message.html.DisplayHtmlFactory
-import com.fsck.k9.ui.changelog.ChangelogViewModel
 import com.fsck.k9.ui.helper.DisplayHtmlUiFactory
 import com.fsck.k9.view.K9WebViewClient
 import com.fsck.k9.view.MessageWebView
@@ -26,6 +25,7 @@ import net.openid.appauth.AppAuthConfiguration
 import net.thunderbird.core.common.mail.html.HtmlSettings
 import net.thunderbird.core.preference.storage.Storage
 import net.thunderbird.feature.account.AccountId
+import net.thunderbird.feature.changelog.internal.ChangelogViewModel
 import net.thunderbird.feature.mail.message.list.ui.dialog.SetupArchiveFolderDialogContract
 import net.thunderbird.feature.mail.message.reader.api.css.CssClassNameProvider
 import net.thunderbird.feature.navigation.changelog.api.ChangeLogMode
