+++ feature/changelog/internal/src/main/kotlin/net/thunderbird/feature/changelog/internal/RecentChangesViewModel.kt
@@ -1,4 +1,4 @@
-package com.fsck.k9.ui.changelog
+package net.thunderbird.feature.changelog.internal
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.asLiveData
