+++ legacy/ui/legacy/build.gradle.kts
@@ -15,6 +15,7 @@ dependencies {
     implementation(projects.mail.common)
     implementation(projects.uiUtils.toolbarBottomSheet)
     implementation(projects.core.android.contact)
+    implementation(projects.feature.changelog.internal)
 
     implementation(projects.core.featureflag)
     implementation(projects.core.logging.api)
