+++ legacy/ui/legacy/src/main/java/com/fsck/k9/UiKoinModules.kt
@@ -8,7 +8,6 @@ import com.fsck.k9.activity.activityModule
 import com.fsck.k9.contacts.contactsModule
 import com.fsck.k9.ui.account.accountUiModule
 import com.fsck.k9.ui.base.uiBaseModule
-import com.fsck.k9.ui.changelog.changelogUiModule
 import com.fsck.k9.ui.choosefolder.chooseFolderUiModule
 import com.fsck.k9.ui.compose.composeModule
 import com.fsck.k9.ui.endtoend.endToEndUiModule
@@ -22,6 +21,7 @@ import com.fsck.k9.ui.messageview.messageViewUiModule
 import com.fsck.k9.ui.settings.settingsUiModule
 import com.fsck.k9.ui.uiModule
 import com.fsck.k9.view.viewModule
+import net.thunderbird.feature.changelog.internal.changelogUiModule
 
 val legacyUiModules = listOf(
     featureAccountOAuthModule,
